test: add coverage thresholds, extract shared test helpers
- Extract duplicated registerVerifyLogin, base32Decode, and generateTotp
  helpers into apps/api/src/__tests__/helpers.ts (was duplicated in 4 files)
- Add coverage thresholds to apps/api Jest config (50% branches/functions,
  60% lines/statements) to prevent coverage regression
- Add coverage thresholds to packages/shared Jest config (70% branches/
  functions, 80% lines/statements)
- Exclude __tests__ directory from API TypeScript compilation to prevent
  test helpers from breaking the production build
- Update vault-crud, admin-dashboard, sessions-activity, and auth-full
  integration tests to import shared helpers

Made-with: Cursor
